+++ backend/canisters/community/api/src/updates/toggle_mute_notifications.rs
@@ -1,15 +1,15 @@
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
-use types::{ReferralType, TimestampMillis};
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
-    pub referral_type: ReferralType,
-    pub codes: Vec<String>,
-    pub expiry: Option<TimestampMillis>,
+    pub mute: bool,
 }
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub enum Response {
     Success,
+    CommunityFrozen,
+    UserNotInCommunity,
+    UserSuspended,
 }