+++ backend/canisters/community/impl/src/updates/change_role.rs
@@ -71,7 +71,11 @@ fn prepare(user_id: UserId, state: &RuntimeState) -> Result<PrepareResult, Respo
                 caller_id: member.user_id,
                 user_index_canister_id: state.data.user_index_canister_id,
                 is_caller_owner: member.role.is_owner(),
-                is_user_owner: state.data.members.get(user_id.into()).map_or(false, |p| p.role.is_owner()),
+                is_user_owner: state
+                    .data
+                    .members
+                    .get_by_user_id(&user_id)
+                    .map_or(false, |p| p.role.is_owner()),
             })
         }
     } else {