+++ backend/canisters/community/impl/src/updates/remove_member.rs
@@ -68,7 +68,7 @@ fn prepare(user_id: UserId, state: &RuntimeState) -> Result<PrepareResult, Respo
             Err(CannotRemoveSelf)
         } else {
             // Check if the caller is authorized to remove the user
-            let is_user_an_owner = match state.data.members.get(user_id.into()) {
+            let is_user_an_owner = match state.data.members.get_by_user_id(&user_id) {
                 None => return Err(TargetUserNotInCommunity),
                 Some(member_to_remove) => {
                     if member