+++ frontend/openchat-client/src/openchat.ts
@@ -170,6 +170,7 @@ import {
     toDatetimeString,
     toLongDateString,
     toShortTimeString,
+    toMonthString,
 } from "./utils/date";
 import formatFileSize from "./utils/fileSize";
 import { calculateMediaDimensions } from "./utils/layout";
@@ -371,8 +372,7 @@ export class OpenChat extends EventTarget {
 
         chatUpdatedStore.subscribe((val) => {
             if (val !== undefined) {
-                const updated = val.updatedEvents;
-                this.chatUpdated(updated);
+                this.chatUpdated(val.chatId, val.updatedEvents);
                 chatUpdatedStore.set(undefined);
             }
         });
@@ -386,7 +386,9 @@ export class OpenChat extends EventTarget {
             .catch((err) => console.error(err));
     }
 
-    private chatUpdated(updatedEvents: UpdatedEvent[]): void {
+    private chatUpdated(chatId: string, updatedEvents: UpdatedEvent[]): void {
+        if (chatId !== this._liveState.selectedChatId) return;
+
         const chat = this._liveState.selectedChat;
         if (chat === undefined) return;
         // The chat summary has been updated which means the latest message may be new
@@ -395,7 +397,10 @@ export class OpenChat extends EventTarget {
             this.handleMessageSentByOther(chat, latestMessage);
         }
 
-        this.refreshUpdatedEvents(chat, updatedEvents);
+        const serverChat = this._liveState.selectedServerChat;
+        if (serverChat !== undefined) {
+            this.refreshUpdatedEvents(serverChat, updatedEvents);
+        }
         this.updateDetails(chat);
         this.dispatchEvent(new ChatUpdated());
     }
@@ -1000,6 +1005,7 @@ export class OpenChat extends EventTarget {
     formatTokens = formatTokens;
     validateTokenInput = validateTokenInput;
     toShortTimeString = toShortTimeString;
+    toMonthString = toMonthString;
     formatMessageDate = formatMessageDate;
     userIdsFromEvents = userIdsFromEvents;
     missingUserIds = missingUserIds;
@@ -3588,6 +3594,7 @@ export class OpenChat extends EventTarget {
                         this.dispatchEvent(new SelectedChatInvalid());
                     } else {
                         chatUpdatedStore.set({
+                            chatId: selectedChatId,
                             updatedEvents: chatsResponse.updatedEvents[selectedChatId] ?? [],
                         });
                     }