+++ backend/canisters/user_index/CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Added
+
+- Add `is_platform_moderator` to `current_user` response ([#3640](https://github.com/open-ic/open-chat/pull/3640))
+- Add `expiry` to referral codes ([#3705](https://github.com/open-ic/open-chat/pull/3705))
+
 ## [[2.0.699](https://github.com/open-ic/open-chat/releases/tag/v2.0.699-user_index)] - 2023-05-18
 
 ### Added